# docs/openapi.yaml
openapi: 3.0.3
info:
  title: RiskLens AI - Enterprise Fraud Intelligence API
  description: |
    Production-grade RESTful API for RiskLens AI. 
    Provides multi-agent fraud investigation, persistent batch transaction ingestion,
    real-time ML risk scoring, Gemini investigator copilot conversations, immutable audit logging,
    and Prometheus/DevOps telemetry.
  version: 2.4.0-prod
  contact:
    name: RiskLens Platform Engineering
    email: platform@risklens.ai
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:3000
    description: Local Container & Development Server
  - url: https://risklens-ai-prod.a.run.app
    description: Google Cloud Run Production Gateway
...
paths:
...
  /api/transactions:
    get:
      tags:
        - Transactions
      summary: Query Transactions
      description: Retrieves transactions from persistent Firestore/in-memory store with filtering.
      operationId: getTransactions
      parameters:
        - name: status
          in: query
          description: Filter by status
          required: false
          schema:
            type: string
            enum: [approved, flagged, held, escalated, rejected, pending]
        - name: tier
          in: query
          description: Filter by risk tier
          required: false
          schema:
            type: string
            enum: [LOW, MEDIUM, HIGH, CRITICAL]
        - name: search
          in: query
          description: Free-text search on transaction ID, customer name, merchant, or location
          required: false
          schema:
            type: string
      responses:
        '200':
          description: List of transactions matching query criteria
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
                  total:
                    type: integer
                    example: 250
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/transactions/{id}:
    get:
      tags:
        - Transactions
      summary: Get Transaction by ID
      description: Retrieves a single transaction record by its unique identifier.
      operationId: getTransactionById
      parameters:
        - name: id
          in: path
          required: true
          description: Transaction identifier (e.g. TX-2026-9041)
          schema:
            type: string
      responses:
        '200':
          description: Transaction details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  responses:
    ValidationError:
      description: Validation Error (400 Bad Request)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardErrorResponse'
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: "Missing 'transactions' field in request body"
            requestId: "req_1723080000_a1b2c3d"
            timestamp: "2026-08-08T00:00:00.000Z"
    NotFoundError:
      description: Resource Not Found (404)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardErrorResponse'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: "Transaction TX-999 not found"
            requestId: "req_1723080000_a1b2c3d"
            timestamp: "2026-08-08T00:00:00.000Z"
...
    InternalServerError:
      description: Internal Server Error (500)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardErrorResponse'
  schemas:
    StandardErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid request parameters
            details:
              type: object
        requestId:
          type: string
          example: req_1723080000_9041xyz
        timestamp:
          type: string
          format: date-time
...
    Transaction:
      type: object
      required:
        - id
        - amount
        - customerId
        - customerName
        - timestamp
      properties:
        id:
          type: string
          example: TX-2026-9041
        amount:
          type: number
          example: 4850.00
        currency:
          type: string
          example: USD
        customerId:
          type: string
          example: CUST-8821
        customerName:
          type: string
          example: Alexander Vance
        merchant:
          type: string
          example: CoinBase Global Exchange
        merchantCategory:
          type: string
          example: Crypto / High Risk
        status:
          type: string
          enum: [approved, flagged, held, escalated, rejected, pending]
          example: flagged
        riskScore:
          type: integer
          example: 87
        riskTier:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
          example: HIGH
        fraudProbability:
          type: number
          example: 0.89
        confidenceScore:
          type: number
          example: 0.94
        location:
          type: string
          example: Lagos, Nigeria
        ipAddress:
          type: string
          example: 102.89.23.11
        deviceType:
          type: string
          example: Linux / TOR Exit Node
        timestamp:
          type: string
          format: date-time
...